requirements.txt
Pillow>=10.0.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
numpy>=1.24.0
PyAutoGUI>=0.9.54
keyboard>=0.13.5 